requirements.txt
python-dotenv
pyodbc
pandas
google-genai